requirements.txt
PyYAML==3.13
selenium==3.14.0
urllib3==1.23
